scripts/requirements.txt
nbformat>=5.9.0
python-dotenv>=1.0.0
rich>=13.0.0
